requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.16
psycopg2==2.9.6
alembic==1.11.1
pydantic==1.10.7
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.1
python-dotenv==1.0.0
